Crate kfs_vi

Visual Compositor

This process takes care of compositing multiple windows on the framebuffer. In the future, it will also be capable of talking to the GPU to provide an OpenGL abstraction layer.

Modules

vbe
VESA Bios Extensions Framebuffer

Structs

Buffer
Internal representation of a window.

IBuffer
IPC Window object

ViInterface
Entry point interface.

Statics

BACKBUFFER_ARR
The backbuffer to draw into.

BUFFERS
A list of the buffers currently alive.

Functions

draw
Draw a portion of a buffer onto the framebuffer.

get_intersect
Gets the intersection between two rectangles.

get_real_bounds
See Buffer::get_real_bounds.

main