Struct kfs_vi::vbe::Framebuffer

pub struct Framebuffer<'a> {
    buf: &'a mut [VBEColor],
    width: usize,
    height: usize,
    bpp: usize,
}

A wrapper around a linear framebuffer. The framebuffer is usually acquired through the map_framebuffer syscall.

Fields
buf: &'a mut [VBEColor]width: usizeheight: usizebpp: usize

Bits-per-pixel. Usually 8.

Methods

impl<'a> Framebuffer<'a>

pub fn new() -> Result<Framebuffer<'static>, Error>

Creates an instance of the linear framebuffer.

Safety

This function should only be called once, to ensure there is only a single mutable reference to the underlying framebuffer.

pub fn new_buffer(
buf: &'a mut [VBEColor],
width: usize,
height: usize,
bpp: usize
) -> Framebuffer<'a>

Creates a backbuffer backed by an in-memory array.

This is useful to avoid flickering and other display artifact. Compositing should happen in such a backbuffer, and the final result should then be copied into the actual framebuffer.

pub fn width(&self) -> usize

framebuffer width in pixels. Does not account for bpp

pub fn height(&self) -> usize

framebuffer height in pixels. Does not account for bpp

pub fn bpp(&self) -> usize

The number of bits that forms a pixel. Used to compute offsets in framebuffer memory to corresponding pixel px_offset = px_nbr * bpp

pub fn get_px_offset(&self, x: usize, y: usize) -> usize

Gets the offset in memory of a pixel based on an x and y.

Panics

Panics if y >= self.height() or x >= self.width()

pub fn write_px(&mut self, offset: usize, color: VBEColor)

Writes a pixel in the framebuffer respecting the bgr pattern

Panics

Panics if offset is invalid

pub fn write_px_at(&mut self, x: usize, y: usize, color: VBEColor)

Writes a pixel in the framebuffer respecting the bgr pattern Computes the offset in the framebuffer from x and y

Panics

Panics if coords are invalid

pub fn get_fb(&mut self) -> &mut [VBEColor]

Gets the underlying framebuffer

pub fn clear(&mut self)

Clears the whole screen

pub fn clear_at(&mut self, x: usize, y: usize, width: usize, height: usize)

Clears a segment of the screen.

Panics

Panics if x + width or y + height falls outside the framebuffer.